Move new methods to corresponding utils
