+++ distributed_training/base/validator.py
@@ -27,80 +27,8 @@ from typing import List, Tuple
 import bittensor as bt
 import numpy as np
 from distributed_training.base.neuron import BaseNeuron
-from distributed_training.utils.uids import get_random_uids
-from distributed_training.validator.reward import get_rewards
+from distributed_training.utils.weight_utils import convert_weights_and_uids_for_emit
 from distributed_training.utils.chain import log_peerid_to_chain
-
-def convert_weights_and_uids_for_emit(
-    uids: np.ndarray, weights: np.ndarray
-) -> Tuple[List[int], List[int]]:
-    r"""Converts weights into integer u32 representation that sum to MAX_INT_WEIGHT.
-    Args:
-        uids (:obj:`np.ndarray,`):
-            Array of uids as destinations for passed weights.
-        weights (:obj:`np.ndarray,`):
-            Array of weights.
-    Returns:
-        weight_uids (List[int]):
-            Uids as a list.
-        weight_vals (List[int]):
-            Weights as a list.
-    """
-    # Checks.
-    uids = np.asarray(uids)
-    weights = np.asarray(weights)
-
-    # Get non-zero weights and corresponding uids
-    non_zero_weights = weights[weights > 0]
-    non_zero_weight_uids = uids[weights > 0]
-
-    # Debugging information
-    bittensor.logging.debug(f"weights: {weights}")
-    bittensor.logging.debug(f"non_zero_weights: {non_zero_weights}")
-    bittensor.logging.debug(f"uids: {uids}")
-    bittensor.logging.debug(f"non_zero_weight_uids: {non_zero_weight_uids}")
-
-    if np.min(weights) < 0:
-        raise ValueError(
-            "Passed weight is negative cannot exist on chain {}".format(
-                weights
-            )
-        )
-    if np.min(uids) < 0:
-        raise ValueError(
-            "Passed uid is negative cannot exist on chain {}".format(uids)
-        )
-    if len(uids) != len(weights):
-        raise ValueError(
-            "Passed weights and uids must have the same length, got {} and {}".format(
-                len(uids), len(weights)
-            )
-        )
-    if np.sum(weights) == 0:
-        bittensor.logging.debug("nothing to set on chain")
-        return [], []  # Nothing to set on chain.
-    else:
-        max_weight = float(np.max(weights))
-        weights = [
-            float(value) / max_weight for value in weights
-        ]  # max-upscale values (max_weight = 1).
-        bittensor.logging.debug(
-            f"setting on chain max: {max_weight} and weights: {weights}"
-        )
-
-    weight_vals = []
-    weight_uids = []
-    for i, (weight_i, uid_i) in enumerate(list(zip(weights, uids))):
-        uint16_val = round(
-            float(weight_i) * int(U16_MAX)
-        )  # convert to int representation.
-
-        # Filter zeros
-        if uint16_val != 0:  # Filter zeros
-            weight_vals.append(uint16_val)
-            weight_uids.append(uid_i)
-    bittensor.logging.debug(f"final params: {weight_uids} : {weight_vals}")
-    return weight_uids, weight_vals
 
 class BaseValidatorNeuron(BaseNeuron):
     """
